Added new functions, started cleaning up the code

+++ Core/Inc/PowerUtils.h
@@ -20,8 +20,8 @@ typedef enum {
   CHARGING_FULL
 } ChargeStatus;
 
-uint32_t      ReadBatteryVoltage(void);   // Returns battery voltage in mV
-uint32_t      ReadSolarVoltage(void);     // Returns solarpanel voltage in mV
+float      ReadBatteryVoltage(void);   // Returns battery voltage in mV
+float      ReadSolarVoltage(void);     // Returns solarpanel voltage in mV
 bool          Check_USB_PowerOn(void);    // True if 5V on USB Connector
 ChargeStatus  Read_Charge_Status(void);   // Returns Chargestatus
